@@ -3,15 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
-    "rxjs": "^5.4.2",
+    "react": "^16.4.0",
+    "react-dom": "^16.4.0",
+    "react-scripts": "1.1.4",
+    "rxjs": "^6.2.0",
+    "rxjs-compat": "^6.2.0",
     "rxsync": "^0.0.3",
     "simple-react-canvas": "^0.0.1",
-    "socket.io-client": "^2.0.3"
-  },
-  "devDependencies": {
-    "react-scripts": "1.0.10"
+    "socket.io-client": "^2.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
